Restore old schema_test for now

+++ tests/schema.py
@@ -11,8 +11,8 @@ import infra.e2e_args
 from loguru import logger as LOG
 
 
-def build_schema_file_path(root, prefix, method):
-    return os.path.join(root, prefix, f"{method}.json")
+def build_schema_file_path(root, method, schema_type):
+    return os.path.join(root, "{}_{}.json".format(method, schema_type))
 
 
 def run(args):
@@ -38,29 +38,38 @@ def run(args):
                 error=lambda status, msg: status == http.HTTPStatus.OK.value,
             )
 
-            target_file = build_schema_file_path(
-                args.schema_dir, client.prefix, method
-            )
             if schema_response.result is not None:
-                schema_found = True
-                formatted_schema = json.dumps(schema_response.result, indent=2)
-                os.makedirs(os.path.dirname(target_file), exist_ok=True)
-                with open(target_file, "a+") as f:
-                    f.seek(0)
-                    previous = f.read()
-                    if previous != formatted_schema:
-                        LOG.debug("Writing schema to {}".format(target_file))
-                        f.truncate(0)
-                        f.seek(0)
-                        f.write(formatted_schema)
-                        changed_files.append(target_file)
+                if len(schema_response.result) != 1:
+                    raise ValueError(
+                        f"This test currently only handles single-verb schema responses - can't handle {method}: {json.dumps(schema_response.result, indent=2)}"
+                    )
+                _, schema_element = schema_response.result.popitem()
+                for schema_type in ["params", "result"]:
+                    element_name = "{}_schema".format(schema_type)
+                    element = schema_element[element_name]
+                    target_file = build_schema_file_path(
+                        args.schema_dir, method, schema_type
+                    )
+                    if element is not None and len(element) != 0:
+                        schema_found = True
+                        formatted_schema = json.dumps(element, indent=2)
+                        os.makedirs(os.path.dirname(target_file), exist_ok=True)
+                        with open(target_file, "a+") as f:
+                            f.seek(0)
+                            previous = f.read()
+                            if previous != formatted_schema:
+                                LOG.debug("Writing schema to {}".format(target_file))
+                                f.truncate(0)
+                                f.seek(0)
+                                f.write(formatted_schema)
+                                changed_files.append(target_file)
+                            else:
+                                LOG.debug("Schema matches in {}".format(target_file))
                     else:
-                        LOG.debug("Schema matches in {}".format(target_file))
-            else:
-                # Ensure there are no out-of-date files for schema which have been removed
-                if os.path.isfile(target_file):
-                    os.remove(target_file)
-                    changed_files.append(target_file)
+                        # Ensure there are no out-of-date files for schema which have been removed
+                        if os.path.isfile(target_file):
+                            os.remove(target_file)
+                            changed_files.append(target_file)
 
             if schema_found:
                 methods_with_schema.add(method)
@@ -80,7 +89,7 @@ def run(args):
             fetch_schema(user_client)
 
         with primary.node_client() as node_client:
-            LOG.warning("node frontend")
+            LOG.info("node frontend")
             fetch_schema(node_client)
 
         with primary.member_client() as member_client:
